arm/dts/connect4: Add I2C6 with eeprom
The I2C6 is used for the eeprom. The eeprom conforms to the HAT
specification for RaspberryPi board. The I2C0 is not intended to be used
as general purpuse I2C by the ARM core. It is reserved for the video
core. But the HAT eeprom needs to be located on the I2C0 GPIO pins 0 and 1.
To access the eeprom at runtime we use the I2C6 on the GPIO pins 0 and
1.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>