arm/dts/connect4: Add LEDs
The Connect4 has one power led which can be set chnaged to red in case
of an error. And two RGB leds on the core board. It has a third RGB led
on the power board.

All leds are connected to the GPIO expanders on the core respectively on
the power board.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>